Use -k to retain output files.

+++ Projects/fluids_color_2d/test-order-one.sh
@@ -10,6 +10,11 @@ else
     shift
 fi
 
+KP=0
+if [ "x$1" == "x-k" ] ; then
+    KP=1
+    shift
+fi
 
 min="$1"
 max="$2"
@@ -22,13 +27,14 @@ for r in `seq $min $max` ; do
     T=`mktemp`
     O=`mktemp -d`
     echo "$((8*$r))" > $T
+    [ $KP = 1 ] && echo OUTPUT $r $O
     K=`$SLAVE -a -o $T -p $r -- "$@" -refine $r -o $O`
     J="$J -d $K"
     OO="$OO $O"
     L="$L $T"
 done
 PPJ=`$SLAVE $J -p 100 -- /bin/bash ./post-process.sh "$out" $L`
-$SLAVE -d $PPJ -p 100 -- /bin/bash -c "rm -r $OO" >/dev/null
+[ $KP = 1 ] || $SLAVE -d $PPJ -p 100 -- /bin/bash -c "rm -r $OO" >/dev/null
 
 if [ $M = 1 ] ; then
     $SLAVE -k
